requirements.txt
numpy
torch
absl_py
accelerate
gin_config
imageio
imageio[ffmpeg]
matplotlib
mediapy
opencv_contrib_python
opencv_python
Pillow
trimesh
pymeshlab
xatlas
plyfile
rawpy
ninja
scipy
scikit-image
scikit-learn
tensorboard
tensorboardX
tqdm
tensorly
